backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
sqlmodel==0.0.14
PyJWT==2.8.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
email-validator==2.0.0
bcrypt==4.0.1
